requirements.txt
# requirements.txt
# Recommendation: Lock versions using 'pip freeze > requirements.txt' or pip-tools after installation

# Core Frameworks
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Langchain & AI
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5 # Needed even if using Ollama via OpenAI compatible API

# Database
mysql-connector-python>=8.3.0
# aiomysql>=0.2.0 # Uncomment if/when migrating db.py to full async driver

# Configuration & Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0 # Required for config.py BaseSettings

# Web & Async HTTP
httpx>=0.25.0 # Required for async report.py upload

# Standard Library Dependencies (usually included, but good practice)
# asyncio # Part of Python standard library
# json # Part of Python standard library
# logging # Part of Python standard library
# platform # Part of Python standard library
# os # Part of Python standard library

# FastAPI Dependency for Form Data (might be needed indirectly)
python-multipart>=0.0.6

# TTS
gTTS>=2.5.1

# Image Processing
Pillow>=9.0.0

# Optional: Add for accurate TTS duration or audio manipulation
# mutagen>=1.45.0
# pydub>=0.25.0